- refactor dice rolling and roll dialog into separate classes - minor improvement to talent display in chat
