+++ .gitignore
@@ -1,1 +1,3 @@
 .idea/
+*.code-workspace
+.history/
