+++ chat/item.html
@@ -115,7 +115,7 @@
         {{#if isTalent}}
         <p><strong>{{localize "TALENT.RANK"}}:</strong> {{data.rank}}</p>
         {{#if data.description}}
-        <p>{{data.description}}</p>
+        <p>{{plaintextToHTML data.description}}</p>
         {{/if}}
         {{/if}}
