+++ script/sheet/character.js
@@ -1,9 +1,6 @@
 import { ForbiddenLandsActorSheet } from "./actor.js";
 
 export class ForbiddenLandsCharacterSheet extends ForbiddenLandsActorSheet {
-  dices = [];
-  lastTestName = "";
-  lastDamage = 0;
 
   static get defaultOptions() {
     return mergeObject(super.defaultOptions, {
@@ -85,7 +82,7 @@ export class ForbiddenLandsCharacterSheet extends ForbiddenLandsActorSheet {
     html.find(".roll-consumable").click((ev) => {
       const consumableName = $(ev.currentTarget).data("consumable");
       const consumable = this.actor.data.data.consumable[consumableName];
-      this.rollConsumable(consumable);
+      this.diceRoller.rollConsumable(consumable);
     });
   }
 
@@ -118,52 +115,6 @@ export class ForbiddenLandsCharacterSheet extends ForbiddenLandsActorSheet {
     this.actor.createEmbeddedEntity("OwnedItem", data, { renderSheet: true });
   }
 
-  push() {
-    this.dices.forEach((dice) => {
-      if ((dice.value < 6 && dice.value > 1 && dice.type !== "skill") || (dice.value < 6 && ["artifact", "skill"].includes(dice.type))) {
-        let die = new Die(dice.face);
-        die.roll(1);
-        dice.value = die.total;
-        let successAndWeight = this.getSuccessAndWeight(dice.value, dice.type);
-        dice.success = successAndWeight.success;
-        dice.weight = successAndWeight.weight;
-      }
-    });
-    this.sendRollToChat(true);
-  }
-
-  async rollConsumable(consumable) {
-    let consumableName = game.i18n.localize(consumable.label);
-    let result;
-    if (!consumable.value) {
-      result = "FAILED";
-    } else {
-      let die = new Die(consumable.value);
-      die.roll(1);
-      if (die.total > 2) {
-        result = "SUCCEED";
-      } else {
-        result = "FAILED";
-      }
-    }
-    let consumableData = {
-      name: consumableName,
-      result: game.i18n.localize(result)
-    }
-    const html = await renderTemplate("systems/forbidden-lands/chat/consumable.html", consumableData);
-    let chatData = {
-      user: game.user._id,
-      rollMode: game.settings.get("core", "rollMode"),
-      content: html,
-    };
-    if (["gmroll", "blindroll"].includes(chatData.rollMode)) {
-      chatData.whisper = ChatMessage.getWhisperIDs("GM");
-    } else if (chatData.rollMode === "selfroll") {
-      chatData.whisper = [game.user];
-    }
-    ChatMessage.create(chatData);
-  }
-
   _getHeaderButtons() {
     let buttons = super._getHeaderButtons();
 
@@ -179,7 +130,7 @@ export class ForbiddenLandsCharacterSheet extends ForbiddenLandsActorSheet {
           label: "Push",
           class: "push-roll",
           icon: "fas fa-skull",
-          onclick: (ev) => this.push(),
+          onclick: (ev) => this.diceRoller.push(),
         },
       ].concat(buttons);
     }
